+++ test/integration/gov/epa/emissions/framework/services/cost/MaxEmsRedStrategyTestCase.java
@@ -155,7 +155,7 @@ public class MaxEmsRedStrategyTestCase extends ServiceTestCase {
         EfficiencyRecord record = new EfficiencyRecord();
         record.setPollutant(pollutant);
         record.setLocale(locale);
-        record.setEfficiency(efficiency);
+        record.setEfficiency((double)efficiency);
         record.setRuleEffectiveness(100);
         record.setRulePenetration(100);
         record.setCostPerTon(cost);
