+++ test/integration/gov/epa/emissions/framework/services/cost/MaxEmsRedStrategyTestDetailedCase.java
@@ -275,7 +275,7 @@ public class MaxEmsRedStrategyTestDetailedCase extends ServiceTestCase {
         EfficiencyRecord record = new EfficiencyRecord();
         record.setPollutant(pollutant);
         record.setLocale(locale);
-        record.setEfficiency(efficiency);
+        record.setEfficiency((double)efficiency);
         record.setRuleEffectiveness(100);
         record.setRulePenetration(100);
         record.setCostPerTon(cost);
